+++ R/addComplement.R
@@ -22,13 +22,20 @@ setMethod("addComplement", signature(x = "transactions"),
   function(x, labels, complementLabels=NULL) {
     
     ### default names for complements
-    if(is.null(complementLabels)) complementLabels <- paste("!", 
-      labels, sep="")
+    if(is.null(complementLabels)) 
+      complementLabels <- paste("!", labels, sep="")
     
+    ### add complements
     add <- sapply(labels, function(y) !(x %in% y))
-    m <- as.matrix(add)
-    colnames(m) <- complementLabels
-    tr <- as(m, "transactions")
+    colnames(add) <- complementLabels
+    tr <- as(add, "transactions")
+    
+    ### add variables and levels to original items
+    if(is.null(itemInfo(x)$variables)) itemInfo(x)$variables <- itemLabels(x)
+    if(is.null(itemInfo(x)$levels)) itemInfo(x)$levels <- factor(rep(TRUE, ncol(x)), levels = c (TRUE, FALSE))
+     
+    itemInfo(tr)$variables <- labels
+    itemInfo(tr)$levels <- factor(rep(FALSE, ncol(tr)), levels = c (TRUE, FALSE))
     
     merge(x, tr)
   })